backend-app/requirements.txt

fastapi==0.104.1

Pillow==10.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
pytz
annotated-types      
anyio                
bcrypt            
certifi             
click              
dnspython           
ecdsa                
email-validator      
exceptiongroup       
fastapi              
greenlet             
h11                
httpcore            
httptools           
httpx                
idna                 
itsdangerous        
Jinja2              
MarkupSafe           
orjson               
passlib              
Pillow               
pip                  
psycopg2-binary      
pyasn1               
pydantic             
pydantic_core        
pydantic-extra-types 
pydantic-settings    
python-dotenv        
python-jose          
python-multipart     
pytz                 
PyYAML               
rsa                  
setuptools          
six                  
sniffio              
SQLAlchemy           
starlette            
typing_extensions    
ujson                
uvicorn              
uvloop               
watchfiles           
websockets           
